Automation rules: Allow decimal input on page per visit (#4988)

+++ client/one/app/features/rules/rules.config.ts
@@ -978,7 +978,7 @@ export const RULE_CONDITIONS_OPTIONS = {
                 ...RULE_CONDITION_OPERANDS_OPTIONS[
                     RuleConditionOperandType.AbsoluteValue
                 ],
-                dataType: DataType.Integer,
+                dataType: DataType.Decimal,
             },
             // RULE_CONDITION_OPERANDS_OPTIONS[
             //     RuleConditionOperandType.PageviewsPerVisitCampaignGoal
